+++ resources/views/classes/table.blade.php
@@ -4,7 +4,9 @@
             <tr>
                 <th>Nom de la classe</th>
         <th>Code de la classe</th>
-                <th colspan="3">Actions</th>
+        @if(Auth::user()->role==1)
+           <th colspan="3">Actions</th>
+           @endif
             </tr>
         </thead>
         <tbody>
@@ -12,6 +14,7 @@
             <tr>
                 <td>{{ $classes->class_name }}</td>
             <td>{{ $classes->class_code }}</td>
+            @if(Auth::user()->role==1)
                 <td>
                     {!! Form::open(['route' => ['classes.destroy', $classes->class_id], 'method' => 'delete']) !!}
                     <div class='btn-group'>
@@ -26,6 +29,7 @@
                     </div>
                     {!! Form::close() !!}
                 </td>
+                @endif
             </tr>
         @endforeach
         </tbody>
